AptVO lat,lng double로 수정

+++ spring_prj/src/main/java/com/kpc/homes/AptVO.java
@@ -24,138 +24,219 @@ public class AptVO {
 	private String aptStr;
 	private String areaStr;
 
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
+
 
 	public long getSeq() {
 		return seq;
 	}
+
+
 	public void setSeq(long seq) {
 		this.seq = seq;
 	}
+
+
 	public String getGu() {
 		return gu;
 	}
+
+
 	public void setGu(String gu) {
 		this.gu = gu;
 	}
+
+
 	public String getDong() {
 		return dong;
 	}
+
+
 	public void setDong(String dong) {
 		this.dong = dong;
 	}
+
+
 	public int getAddr1() {
 		return addr1;
 	}
+
+
 	public void setAddr1(int addr1) {
 		this.addr1 = addr1;
 	}
+
+
 	public int getAddr2() {
 		return addr2;
 	}
+
+
 	public void setAddr2(int addr2) {
 		this.addr2 = addr2;
 	}
+
+
 	public String getApt() {
 		return apt;
 	}
+
+
 	public void setApt(String apt) {
 		this.apt = apt;
 	}
+
+
 	public String getGubun() {
 		return gubun;
 	}
+
+
 	public void setGubun(String gubun) {
 		this.gubun = gubun;
 	}
+
+
 	public String getArea() {
 		return area;
 	}
+
+
 	public void setArea(String area) {
 		this.area = area;
 	}
+
+
 	public int getConym() {
 		return conym;
 	}
+
+
 	public void setConym(int conym) {
 		this.conym = conym;
 	}
+
+
 	public int getPrice() {
 		return price;
 	}
+
+
 	public void setPrice(int price) {
 		this.price = price;
 	}
+
+
 	public int getDeposit() {
 		return deposit;
 	}
+
+
 	public void setDeposit(int deposit) {
 		this.deposit = deposit;
 	}
+
+
 	public int getMrent() {
 		return mrent;
 	}
+
+
 	public void setMrent(int mrent) {
 		this.mrent = mrent;
 	}
+
+
 	public int getFloor() {
 		return floor;
 	}
+
+
 	public void setFloor(int floor) {
 		this.floor = floor;
 	}
+
+
 	public int getArchyear() {
 		return archyear;
 	}
+
+
 	public void setArchyear(int archyear) {
 		this.archyear = archyear;
 	}
+
+
 	public String getStreet() {
 		return street;
 	}
+
+
 	public void setStreet(String street) {
 		this.street = street;
 	}
+
+
 	public String getSearchKey() {
 		return searchKey;
 	}
+
+
 	public void setSearchKey(String searchKey) {
 		this.searchKey = searchKey;
 	}
+
+
 	public String getSearchStr() {
 		return searchStr;
 	}
+
+
 	public void setSearchStr(String searchStr) {
 		this.searchStr = searchStr;
 	}
+
+
 	public String getAptStr() {
 		return aptStr;
 	}
+
+
 	public void setAptStr(String aptStr) {
 		this.aptStr = aptStr;
 	}
+
+
 	public String getAreaStr() {
 		return areaStr;
 	}
+
+
 	public void setAreaStr(String areaStr) {
 		this.areaStr = areaStr;
 	}
 
-	public int getLat() {
+
+	public double getLat() {
 		return lat;
 	}
-	public void setLat(int lat) {
+
+
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
-	public int getLng() {
+
+
+	public double getLng() {
 		return lng;
 	}
-	public void setLng(int lng) {
+
+
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
-	
-	
+
 
 	@Override
 	public String toString() {
